+++ handler.go
@@ -2,10 +2,10 @@ package healthz
 
 import "net/http"
 
-// NewHealthServiceHandler creates an http.Handler from user configured Probes
+// NewHealthServiceHandler creates an http.Handler from user configured Checkerss
 // The handler is a standard http.ServeMux
-func NewHealthServiceHandler(livenessProbe *Probe, readinessProbe *Probe) http.Handler {
-	healthService := NewHealthService(livenessProbe, readinessProbe)
+func NewHealthServiceHandler(livenessChecker Checker, readinessChecker Checker) http.Handler {
+	healthService := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
 
 	mux.HandleFunc("/healthz", healthService.HealthStatus)
@@ -16,18 +16,18 @@ func NewHealthServiceHandler(livenessProbe *Probe, readinessProbe *Probe) http.H
 
 // HealthStatus checks if the application is healthy
 func (s *HealthService) HealthStatus(w http.ResponseWriter, r *http.Request) {
-	s.checkStatus(w, r, s.livenessProbe)
+	s.checkStatus(w, r, s.livenessChecker)
 }
 
 // ReadinessStatus checks if the app is ready for accepting request (eg. database is available as well)
 func (s *HealthService) ReadinessStatus(w http.ResponseWriter, r *http.Request) {
-	s.checkStatus(w, r, s.readinessProbe)
+	s.checkStatus(w, r, s.readinessChecker)
 }
 
-// Since both health check rely on probes, common logic for them is here
-func (s *HealthService) checkStatus(w http.ResponseWriter, r *http.Request, p *Probe) {
+// Since both health check rely on Checkers, common logic for them is here
+func (s *HealthService) checkStatus(w http.ResponseWriter, r *http.Request, c Checker) {
 	// If the probe fails, we return an error
-	if err := p.Check(); err != nil {
+	if err := c.Check(); err != nil {
 		w.WriteHeader(http.StatusServiceUnavailable)
 		w.Write([]byte("error"))