+++ handler_test.go
@@ -11,12 +11,12 @@ import (
 
 func TestHealthService_HealthStatus(t *testing.T) {
 	checker := new(CheckerMock)
-	livenessProbe := NewProbe(checker)
-	readinessProbe := new(Probe)
+	livenessChecker := NewCheckers(checker)
+	readinessChecker := new(Checkers)
 
 	checker.On("Check").Return(nil)
 
-	service := NewHealthService(livenessProbe, readinessProbe)
+	service := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
 	mux.HandleFunc("/healthz", service.HealthStatus)
 
@@ -31,12 +31,12 @@ func TestHealthService_HealthStatus(t *testing.T) {
 
 func TestHealthService_HealthStatus_Fail(t *testing.T) {
 	checker := new(CheckerMock)
-	livenessProbe := NewProbe(checker)
-	readinessProbe := new(Probe)
+	livenessChecker := NewCheckers(checker)
+	readinessChecker := new(Checkers)
 
 	checker.On("Check").Return(ErrCheckFailed)
 
-	service := NewHealthService(livenessProbe, readinessProbe)
+	service := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
 	mux.HandleFunc("/healthz", service.HealthStatus)
 
@@ -51,12 +51,12 @@ func TestHealthService_HealthStatus_Fail(t *testing.T) {
 
 func TestHealthService_ReadinessStatus(t *testing.T) {
 	checker := new(CheckerMock)
-	livenessProbe := new(Probe)
-	readinessProbe := NewProbe(checker)
+	livenessChecker := new(Checkers)
+	readinessChecker := NewCheckers(checker)
 
 	checker.On("Check").Return(nil)
 
-	service := NewHealthService(livenessProbe, readinessProbe)
+	service := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
 	mux.HandleFunc("/readiness", service.ReadinessStatus)
 
@@ -71,12 +71,12 @@ func TestHealthService_ReadinessStatus(t *testing.T) {
 
 func TestHealthService_ReadinessStatus_Fail(t *testing.T) {
 	checker := new(CheckerMock)
-	livenessProbe := new(Probe)
-	readinessProbe := NewProbe(checker)
+	livenessChecker := new(Checkers)
+	readinessChecker := NewCheckers(checker)
 
 	checker.On("Check").Return(ErrCheckFailed)
 
-	service := NewHealthService(livenessProbe, readinessProbe)
+	service := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
 	mux.HandleFunc("/readiness", service.ReadinessStatus)