+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - `HealthService.RegisterHandlers` method
 - `Type` method from `Checker` interface
+- `Probe` in favor of `Checker` interface and collection
 
 
 ## 0.3.0 - 2017-02-09